Refactor input and suggest in two different components.
