+++ static/js/shared/util.ts
@@ -178,3 +178,16 @@ export function removeSpinner(containerId: string): void {
     }
   }
 }
+
+
+export function stripPatternFromQuery(
+  query: string,
+  pattern: string
+): string {
+  const regex = new RegExp("(?:.(?!" + pattern + "))+([,;\\s])?$", "i");
+
+  // Returns the query without part that matched a result.
+  // E.g.: query: "population of Calif", matched_query: "Calif",
+  // returns "population of "
+  return query.replace(regex, "");
+}
